reformat: Naming et séparation fichier règles métiers
